Also race the fallback, with 5, and no duplicates
